Finished get_auth_code, renamed auth_code_valid and added to AuthCode
Realised needed to keep track of what scope the client was
requesting.  To address this added a scope field to the AuthCode
model.

Realised that having a separate validation function for auth code was
not necessary.  Renamed the auth_code_valid function to get_auth_code,
having it return the AuthCode model from database if valid and False
if something is suspect like its nonexistence in the database.

Finally finished the create and return authorisation code part of
OAuth 2.0.  Next step, return access token on request with
authorisation token.  Will also do the renew auth token with refresh
token.