Added another test, fixed bugs that were detected as a result
Decided to have refresh token reference to the access token be the
code string and not a reference to the token class itself.  This
simplifies things and makes a decision which more importantly allows
me to get on with things.

When using the functions it is important that where the database is
accessed other functions which also open an access to the db are also
not used between the opening and closing of the db.

I am starting to think I should be adding a db parameter to pass a
already open db object to the function.