Fixed redirect bug. Made it possible to get access token without login
HTTPRedirect is a exception, and can't be returned as I previously had
it doing.  replaced 'return' with 'raise' to fix that problem.

I ran into a problem that the access token request was asking for a
username password via session login.  This is not desirable due to the
nature of the source of the request.  The client program does not want
to be authenticating itself twice, once for accessing the access token
url and second time in the actual request itself, according to OAuth
2.0 protocol.  Fixed this by adding the anonymous method and having
that check for the request path.  If request path is the access token
path, return 'anonymous' as the username which is assigned to
cherrypy.request.login.  This is a hack, and will be changed in the
future when I choose a different authentication scheme for cherrypy.

Found out that get_auth_token is a function defined in both
authenticate module and database module.  I was thinking I was using
the one form database module, but since tha function is in the
authenicate module, was using that one instead.  Caused errors as
parameters are different.  Fixed it by importing the one from database
as db_get_auth_token.  Am thinking will just import database and call
functions from it using the dot operator in future.  For now it can
stay as is.