Realised I had not implemented Implicit Grant properly, so fixed it
It now correctly returns an access token and expire_in minus the
refresh token as part of the redirect_uri in the form of a url
fragment. ie http://blah#access_token=j0988k&expires_in=3600.  Forgot
to add the token_type=bearer tag. Will do that for next commit.