Tried to simplify code and avoid DB conflicts, and fixed a few bugs
Tried to get it so the creation of the DB object was covered by the
try/catch/finally block in each function that uses it.  I tried to stick
to the structure of putting all DB creation of, access of, and writes to
DB were in the try block.  All exceptions should log to the error log,
and the message should log what function was being used.  The finally
block should always be used to close the DB connection.  All successful
function returns should happen in the try block where the objects
returned are created.  If a failure/error happens it should let the
function flow fall through to the end of the function where it should
return None or False.

def function():
  try:
    #DB stuff here
    db - DB(...)

    return <success (either data sort or True)>
  except Exception, e:
    #log to error log here
    logging.error('<function name>: <error msg>')
  finally:
    #close db connection here
    db.close()

  return <fail (False or None)>