Realised my mistake and put the DB object creation outside the try/except block
if a exception happens on the DB object creation the finally block
will try and do a close on a non existent DB object.  This will cause
more problems.  I still need to figure out if I want to do a
try/except block around the DB object creation.  I feel I won't die to
the extra layer of complexity it will introduce and the fact that it
will make the code a little harder to read.  If it fails the entire
OAuth 2.0 key provider will not work anyway, so letting the exception
get handled as a 500 HTTP response is probably best anyway.