Created Association class, added new exceptions and started cleaning up, finished off associate_client_with_user
In models.py created the Association class which will be used to store
clients which a user has allowed access to resources for.

associate_client_with_user is finished.  It uses the Association class
to associate a user with clients allowed access to their resources. It
is to be stored in the ZODB database using
''.join(['client_association_', str(user.id)]) as the key.

Exceptions were created to be raised when a user already exists in the
add_user function, and the same for association of clients when they
are already there.  This is done as a new association should not have
the same id as a previous association.