Finished optimising string concatenation and added tests for add_user and associate_client_with_user
String concatenation was being done using the string format method of
'%s' % (some_string).  This is not optimal, and could cause
performance issues in a demanding environment where lots of calls are
made.  The attempted remedy was to use ''.join(some_list) ie
''.join(['one: ', str(variable), ' three']).  The other option would
be to use StringIO, but it seems it only outperforms the chosen
solution if the number of items to concatenate is huge.