Updated associate_client_with_user function and modified Provider class
associate_client_with_user function now checks to see if the client is
of the type 'confidential'.  This type of client should never get a
refresh token, and thus has to reauthorise ever time the access token
expires.  This makes the association redundant, since it's not going
to last anyway.

Am starting to create the interface for creating clients.  Am also
considering where to put the association clients with users function
calls.  I want to make this code maintainable in the future, so just
can't put function calls anywhere.