Created new functions and modified others.
Added a call to database.associate_client_with_user in
authenticate.process_auth_code_grant.  Modified associate function to
take a refresh_token and removed client as refresh_token is a instance
of RefreshToken and contains an instance of User and Client.  Will
remove user from the parameters for this function since it's in
refresh_token anyway.

Created database.get_associations since this is going to be needed
later on, as well as starting database.delete_user.

There were some other modifications like, making ''.join([str, str])
into str + str.  This is because I've found that using str + str is
the fastest concatenation method in Python 2.7, unlike previous
versions where it was the slowest.

This database module code is going to need revising and reorganised,
due to the fact that it is becoming hard to find functions, and there
are functions that go together.  It will make it easier to maintain
later too.