Modified users.py
Removed default value for get_user(), you now have to supply a uid.
This is and effort to remove the dependency on cherrypy.

Also made it so users of uid of None can't be created.

Added more validity checks and outputted more useful log messages,
using logging.  Made sure that the string concatenation can take place
even if the var is None, by using the str(var) function.