Fixed db.update bug, create a global to keep db open
The problem for db.update came from the fact that in token.get_token
it did a deepcopy on the token object.  This made it a copy of that
object, but without the connection to the DB.  When trying to update
that object, the original was still unchanged, and marked as such.  It
did not update as a result.  The solution was to remove the deepcopy
part and return the original object, which is still tied to the
database.  So when it's updated, the update will stick.

I noticed a bit of a performance issue when doing API calls.  Doing a
bit of research I found one issue was the creation of new database
connections for every query.  This is inefficient, so I created a hack
which made some global variables in the DB module.  These repesented
the connection to the database, which when a new ZODB object was
created, if these globals were set, it would use these to set its
attributes, otherwise it would create a new connection, and set the
globals with these values.  I noticed an immediate improvement in
response times.  I need to investigate creating a better Singleton
design pattern for the ZODB, if there is one that is worth changing my
current solution for.