Fixed a error, made ZODB a singleton, update_association change
Toke flow now passes the users uid to
create_implicit_grant_access_token(...), which it didn't before thus
failed and returned False.

Created a singleton decorator and used it to decorate ZODB.  So
whenever a new ZODB() instance is attempted, it will create one if one
doesn't exist, or return the already created one if it does.  This
means there will be no more than one ZODB instance created at one
time.  In most of my modules I import this class as 'from DB import
ZODB as DB'.

Tidied up ZODB getting rid of logging calls and the commit call in
update.

Now when a association is updated, the old access and refresh tokens
get deleted from the database.  I like my functions and methods to
clean up after themselves.

In ZODB.delete, the check for isinstance AuthCode has been changed to
check to see if it is an instance of Code.  AuthCode does not cover
enough of the models used for token storage.