Added README file, modified DB.py and client.py
Added a README file, so that people will get a better understanding of
what this project is all about.  This also complies with what Github
keeps asking for.

In DB.py removed the unnecessary connection pool decorator.

In client.py stopped create_client from making deepcopy of the client
record.  It's not necessary and slows down the execution of code.